src/1-ingestion/requirements.txt
# ------------------------------------ app\3-ingestion\requirements.txt ------------------------------------ 
requests==2.31.0
azure-identity==1.12.0
azure-storage-file-datalake==12.13.0
tenacity==8.2.3